feat(stdlib): move addDuration and subDuration from experimental to date
