+++ libflux/go/libflux/buildinfo.gen.go
@@ -109,8 +109,9 @@ var sourceHashes = map[string]string{
 	"stdlib/contrib/tomhollingworth/events/duration_with_stop_test.flux":                          "ec06b590cf7010b89e57f1bb3aee7374fb46290dc8d3f441f5d12d94bce47925",
 	"stdlib/csv/csv.flux":                                                                         "94a1d8dd59c0e092617e9c974bad4edaf1a2c6ebb20fd185524b58cb657329b0",
 	"stdlib/csv/csv_test.flux":                                                                    "8aaf56893a1aefb65883ba96ebdb0283670d76ec872bfedfe389d778bdc1bb8f",
-	"stdlib/date/date.flux":                                                                       "2a898854a3e1aac514bb042eee530667ce6d32ee9a8bc3ea07a57ebf72f4e14c",
+	"stdlib/date/date.flux":                                                                       "e69ab4edc0d04b25cc29d1ce68374dd560584f747cce18b96e93e2518e665de6",
 	"stdlib/date/date_test.flux":                                                                  "42ea61ee595ab6cfc5631505cf5a757f5a7fa24bd5cc43466099956552b071cf",
+	"stdlib/date/durations_test.flux":                                                             "07d63e69eaf84964118a7d5165cb64b659fe1cc39a9a914dbfd1e704cac56201",
 	"stdlib/date/hour_duration_test.flux":                                                         "7da9263cddf0fa443e9dc55b176139d1dd67cb689c0be243d8b595799fc066f4",
 	"stdlib/date/hour_time_test.flux":                                                             "23007876c06a11d25f62368aa1ad6f66ed230dd917deac912a390ec8f72d7fc5",
 	"stdlib/date/microsecond_duration_test.flux":                                                  "d3fc3aaf481a6c52dc94b7cbfad4ca57b07f50535e96d76f4db8497fcb9ecbb0",
@@ -158,7 +159,7 @@ var sourceHashes = map[string]string{
 	"stdlib/experimental/csv/csv.flux":                                                            "94ae72b5fb50b8d65e9b384aebaa890f04bc914bb344ea751a64f729b5c849fe",
 	"stdlib/experimental/distinct_test.flux":                                                      "ebc7f2bf1bc49266a8809f8a3e5cffe5467ac2774c14d40e50c85d14f8c6cd87",
 	"stdlib/experimental/durations_test.flux":                                                     "6394bbfc88ab51a42cfee4a0f1e6ed327b860662922dd99aa6f99ebda5b167ee",
-	"stdlib/experimental/experimental.flux":                                                       "30a8b0ca7bda907990c87b641a06684b824ff281f9318fa7cb33e13263ac39ac",
+	"stdlib/experimental/experimental.flux":                                                       "2c057d77314fbee46fbc46038029d5ae02776585be7d593fd50d7eff33e33f97",
 	"stdlib/experimental/experimental_test.flux":                                                  "206cbb54d35da80a6736dec990219f5bbabcdc3c65fbb9d1041058c8aae3450d",
 	"stdlib/experimental/fill_test.flux":                                                          "adc4f95f746ccdaf4a0843633753e723a4524ea0afbfb2cd6b70726d855ebfb8",
 	"stdlib/experimental/first_test.flux":                                                         "3dffabf42ff18aa9a6515010e3496f13a44e0429c38636f6962361dbc8693c3f",
