+++ stdlib/date/durations.go
@@ -1,31 +1,27 @@
-package experimental
+package date
 
 import (
 	"context"
 	"fmt"
 
-	"github.com/influxdata/flux/codes"
 	"github.com/influxdata/flux/execute"
 	"github.com/influxdata/flux/internal/date"
-	"github.com/influxdata/flux/internal/errors"
-	"github.com/influxdata/flux/interpreter"
 	"github.com/influxdata/flux/runtime"
-	"github.com/influxdata/flux/semantic"
 	"github.com/influxdata/flux/values"
 )
 
 const (
-	addDurationTo        = "_addDuration"
-	subtractDurationFrom = "_subDuration"
+	addDurationTo        = "_add"
+	subtractDurationFrom = "_sub"
 )
 
 func init() {
-	runtime.RegisterPackageValue("experimental", addDurationTo, addDuration(addDurationTo))
-	runtime.RegisterPackageValue("experimental", subtractDurationFrom, subDuration(subtractDurationFrom))
+	runtime.RegisterPackageValue("date", addDurationTo, addDuration(addDurationTo))
+	runtime.RegisterPackageValue("date", subtractDurationFrom, subDuration(subtractDurationFrom))
 }
 
 func addDuration(name string) values.Value {
-	tp := runtime.MustLookupBuiltinType("experimental", "addDuration")
+	tp := runtime.MustLookupBuiltinType("date", "add")
 	fn := func(ctx context.Context, args values.Object) (values.Value, error) {
 		d, ok := args.Get("d")
 		if !ok {
@@ -54,7 +50,7 @@ func addDuration(name string) values.Value {
 }
 
 func subDuration(name string) values.Value {
-	tp := runtime.MustLookupBuiltinType("experimental", "subDuration")
+	tp := runtime.MustLookupBuiltinType("date", "sub")
 	fn := func(ctx context.Context, args values.Object) (values.Value, error) {
 		d, ok := args.Get("d")
 		if !ok {
@@ -80,31 +76,4 @@ func subDuration(name string) values.Value {
 		return values.NewTime(lTime.Time().Add(d.Duration().Mul(-1))), nil
 	}
 	return values.NewFunction(name, tp, fn, false)
-}
-
-func getLocation(args values.Object) (string, values.Duration, error) {
-	var name, offset values.Value
-	var ok bool
-	a := interpreter.NewArguments(args)
-	if location, err := a.GetRequiredObject("location"); err != nil {
-		return "UTC", values.ConvertDurationNsecs(0), err
-	} else {
-		name, ok = location.Get("zone")
-		if !ok {
-			return "UTC", values.ConvertDurationNsecs(0), errors.New(codes.Invalid, "zone property missing from location record")
-		} else if got := name.Type().Nature(); got != semantic.String {
-			return "UTC", values.ConvertDurationNsecs(0), errors.Newf(codes.Invalid, "zone property for location must be of type %s, got %s", semantic.String, got)
-		}
-
-		if offset, ok = location.Get("offset"); ok {
-			if got := offset.Type().Nature(); got != semantic.Duration {
-				return "UTC", values.ConvertDurationNsecs(0), errors.Newf(codes.Invalid, "offset property for location must be of type %s, got %s", semantic.Duration, got)
-			}
-		}
-	}
-	if name.IsNull() {
-		name = values.NewString("UTC")
-	}
-
-	return name.Str(), offset.Duration(), nil
 }
